Mejora
Se implementa la correccion del angulo para que se mantenga en el suelo, se realiza el cambio de herramienta y se mejora la interaccion con la interfaz, a falta de implementar el codigo que permita leer los comandos.

+++ Motor.cpp
@@ -41,7 +41,7 @@ float motor::getFeedback(){
   }
 }
 
-void motor::setVelocidad(float a){
+void motor::setVelocidad(int a){
   velocidad = a;
   analogWrite(pin_pwm, velocidad);
 }
